backend/requirements.txt
django==5.1
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dotenv==0.9.9
psutil==5.9.0
psycopg2-binary==2.9.9
qrcode==8.2
pillow==10.4.0